requirements.txt
crewai
langchain-groq
langchain-community
sqlalchemy
psycopg2-binary
python-dotenv
langchain-ollama